Fix a couple bugs
One had to do with broken animation when ‘painting’ cells in edit mode.

The other was about filling blank rows after reading in pattern files.
I think this only occurs in special cases (like inaccurate width/height
given in the file), and the way we measure the difference in height
different from how we were doing it before.